Project 4 Optimization of marketing expenses.
We started our analysis by calculating the number of people that use our product on a daily/weekly/monthly basis. Using plots we have noticed an uptrend that begins in a mid year and reaches it's peak in December. We confirmed this by plotting the metrics on 3 graphs that represent 3 different timeframes (daily, weekly, monthly).

Comparing the "touch" vs "desktop" devices across the time period we discovered that the desktop is by far a more popular device for our product. Is it because our mobile app is not developed enough? We have also shown on a graph that both devices follow the same overall ups and downs buth differ on the magnitude of values (visits).

We then calculated sessions per day and found that on avr the number of daily sessions per user is 1.08 and it is normally distributed in both directions - we confirmed this visually by a graph. Our sessions per user, doesn't rise and keeps bouncing around the 1.08, this is not a good sign. We want people to use our service multiple times a day.

We then calculated the length of each session and found that the ASL mode is 60 seconds long. We also delivered a few possible explanation why this might be the case.

We continued with calculating the retention rate. We created cohorts based on the first visit of the user and showed on a heatmap that although retention in different cohorts are not the same, they share the same tendencies: as the time goes by the retention drops. The first cohorts (months 6,8,9) demonstrate some very impressive retention rates in the first third of the period. These higher retention rates could be explained by seasonality, better marketing or changes in the overall market (exit of competitors). At this stage we can not provide a valid explanation for higher retention rates.

Convertion (in days) graph showed that the vast majority of users convert to buyers in the first few days. We also concluded that source_id 1,2, and 5 produce a better conversion than sources 9,10, and 4.

AVR Revenue by month showed us that the revs are in uptrend in the second half of the year and peak during December and then sharply drop. The same exact dinamics is noticed in the marketing total expenditure.

The difference between cohorts continues in avr purchase size metrics. If we plot the heatmap by cohorts we will see some cohorts (Sept2017 and Dec 2017) are especially active. Their avr purchase size is very high and remains so for a long time. Sept Cohort is also had the highest ltv. Is it because of the marketing?

When we calculate the marketing expenditures we see that the marketing budget was not devided equaly across the sources. Ad source_id number 3 is by far the most popular… (Facebook? Google ads?) But when we calculate the CAC by ad source_id - then this source is also has the highest CAC value… It is the most expensive ad source... Perhaps we should not spend so much on source_id number 3… We confirm this, once again, by looking on ROMI by source_id plot. Seems like source_id number 3 has the LOWEST romi. Perhaps we should invest more in the source_id number 1, as it has the highest ROMI value.

Considering the ROI by cohorts, the July, Sept and Feb cohorts have higher ROI than any other cohort. Our bussiness strategy in these cohorts should be closely inspected and duplicated if we aim to get a faster and higher ROI in other cohorts.

Moving to sales we found that not all source_id are equal. Some sources are “pushing” the users faster towards the conversion, for other it takes a longer time. Source 5, 1 and 2 demonstrate the shorter conversion period and we should take this into account when we structure our marketing efforts. Dividing the sales by cohorts we see that some cohorts are dominate the heatmap. Aug, Sept and Nov cohorts have significantly higher orders by buyer values. And sources 1,2 and 9 have the highest return on marketing investment.