Project 6 Marker Research - Data Storytelling
We’ve decided to open a small robot-run cafe in Los Angeles. The project is promising but expensive, so our partners decide to try to attract investors. They’re interested in the current market conditions—will you be able to maintain your success when the novelty of robot waiters wears off?
Our goal is to prepare some market research in LA.