Integrated Project 2 ML Model for oil industry
Task: Our company develops efficiency solutions for heavy industry. The ML model should predict the amount of gold recovered from gold ore. We have the data on extraction and purification. The model will help to optimize production and eliminate unprofitable parameters.

Analyzing the data of Gold ore purification we showed that the various processing stages have a distinct impact on the concentration of gold, silver, and lead in the raw material as it moves from one stage to the next. Using MAE as a metric, we calculated the Au recovery and showed that it was identical to the true recovery metrics in the data. The datasets required some preprocessing and cleaning: our calculations and models couldn't tolerate division by zero, therefore we had to drop the instances of both Nans and zeros in the target columns. We then created two prediction models: Linear Reg and Forest Reg and calculated the result on our leading metric called "final sMAPE", a variation of MAE. We then compared our result to a dummy regressor that always predicts the mean of the dataset. The forest regression model performed worse than the dummy regressor but the Linear regressor showed a better result.