12 Time Series
Lift Taxi company has collected historical data on taxi orders at airports. To attract more drivers during peak hours, we need to predict the amount of taxi orders for the next hour. Build a model for such a prediction.
The RMSE metric on the test set should not be more than 48.
Project instructions
Download the data and resample it by one hour.
Analyze the data.
Train different models with different hyperparameters. The test sample should be 10% of the initial dataset.
Test the data using the test sample and provide a conclusion.